implementation for service, mySQL added

+++ taxiservice/src/main/java/com/solvd/taxiservice/db/dao/IDAO.java
@@ -3,7 +3,7 @@ package com.solvd.taxiservice.db.dao;
 public interface IDAO<T>{
 
     void create(T t);
-    T getById(int id);
+    T getById(long id);
     void update(T t);
     void delete(T t);
 }
